+++ src/Atc.Cosmos.EventStore/Streams/IStreamSubscriptionFactory.cs
@@ -4,7 +4,6 @@ internal interface IStreamSubscriptionFactory
 {
     IStreamSubscription Create(
         ConsumerGroup consumerGroup,
-        SubscriptionStartOptions startOptions,
         ProcessEventsHandler eventsHandler,
         ProcessExceptionHandler exceptionHandler);
 }