+++ CHANGELOG.md
@@ -9,7 +9,8 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ### Added
 
--   Introduce hard limits to the number of events the system can accept per operation. 
+- Introduced options to provide events async when testing commands using `ICommandGiven`, `ICommandWhen` and `ICommandThen`.
+- Introduce hard limits to the number of events the system can accept per operation. 
     - A maximum of 10 events per command context (CQRS)
     - A maximum of 50 events per stream batch (Event Store)